aider: Extend the `executeAction` function to use `typeText` in the `INPUT` case.
# Aider chat conversation:

ASSISTANT: To extend the `executeAction` function to use `typeText` in the `INPUT` case, we need to modify `clickolas-cage\src\contentScript\index.js`.

Here is the *SEARCH/REPLACE* block:

clickolas-cage\src\contentScript\index.js
```javascript
<<<<<<< SEARCH
    case 'INPUT':
      console.log(`Inputting text: ${param} into field with label: ${label}`)
      document.querySelector(selector).value = param
      await waitForWindowLoad()
      return true
=======
    case 'INPUT':
      console.log(`Inputting text: ${param} into field with label: ${label}`)
      const inputElement = document.querySelector(selector);
      inputElement.focus();
      await typeText(param);
      inputElement.blur();
      await waitForWindowLoad();
      return true;
>>>>>>> REPLACE
```

This change will replace the existing `INPUT` case in the `executeAction` function. Instead of directly setting the value of the input field, it will now use the `typeText` function to simulate typing the text into the input field.