aider: Create a new helper function `runPressTabInTabWithNextStep` and update `handleRequest` to use it.
# Aider chat conversation:

USER: make a new helper function for when we call `runFunctionXTimesWithDelay` with `pressTabInTab`. we need to send the following two lines including `sendMessageToBackgroundScript` with 'next_step' AFTER all the x times with delay are finished. though we can not use await in the request handler as it will close the websocket with backgroundjs

ASSISTANT: To implement the new helper function, we will:

1. Create a new helper function `runPressTabInTabWithNextStep`.
2. Modify the `handleRequest` function to use this new helper function.

Here are the *SEARCH/REPLACE* blocks:

clickolas-cage/src/sidepanel/SidePanel.jsx
```javascript
<<<<<<< SEARCH
  const handleRequest = async (request, sender, sendResponse) => {
=======
  const runPressTabInTabWithNextStep = (times, delay) => {
    runFunctionXTimesWithDelay(pressTabInTab, times, delay).then(() => {
      sendMessageToBackgroundScript({
        type: 'next_step',
      });
    });
  };

  const handleRequest = async (request, sender, sendResponse) => {
>>>>>>> REPLACE
```

clickolas-cage/src/sidepanel/SidePanel.jsx
```javascript
<<<<<<< SEARCH
        runFunctionXTimesWithDelay(pressTabInTab, 10, 250)

        // setOriginalPlan(request.originalPlan)
        // sendMessageToBackgroundScript({
        //   type: 'next_step',
        // })
=======
        runPressTabInTabWithNextStep(10, 250);
>>>>>>> REPLACE
```

This will ensure that the `sendMessageToBackgroundScript` with `type: 'next_step'` is called after all the `pressTabInTab` calls are completed.